SITES-618 improve editorial index performance
- Loading the Section now preloads the home_node for the Section
- ApplicationController#current_user now preloads all of the User roles

Also added a feature to ensure that Section#name is updated when
SectionHome#name is updated. Ideally we can restore the delegate at some
point but Sections are created *before* the SectionHome in the admin
area and they will need to display a name in the meantime.

Fixed related breakages.