SITES-586 delete the demo spec and database dump loader
The database restore for the demo spec was causing referential integrity
issues on restore since adding foreign key checks in the preious commit.

This is because in pgdumps, the order of the tables is not insertion
order but it doesn't matter because constraints are enabled at
the end of the script.

In the case of the demo script, the database will already be present
(but empty of data) with constraints enabled and the COPY blocks in the
dump fail to apply due to key violations.

It turns out that since the demo script is no longer something we need
to maintain it is simpler to delete it than engineer around this
problem.