SITES-656 redirect authentication failures to /users/sign_in
Authentication behaviour is now handled by using the AccessDeniedConcern
in appropriate controllers. This centralises the behaviour for
consitency. The behaviour is:

- When a user is not authenticated:

  If they try to access a page that requires authorisation they will be
  redirected to /users/sign_in

  If it's a public page the request will succeed

- When a user is authenticated:

  If the user is also authorised the request will succeed.
  If they are not authorised the request will fail with a 401.

The flash messages are also set by the helper.

I deleted some of the specs for various controllers that were testing
the very rudimentary behaviour of how authentication and authorisation
failures were handled.  The basics are now tested in the spec for the
AccessDeniedConcern itself so I removed the duplication.