SITES-701 set Cache-Control headers
Use Cache-Control headers with a max age of 5 minutes and set to
publicly cacheable to supplement the ETAG system that was already in
place.

In addition, simplify the cacheing API on ApplicationController. Instead
of `bustable_fresh_when` and `bustable_stale?`, have a single method
called `with_caching` that takes the same arguments as the old methods
in addition to a block argument. The form with the block argument is
used in all situation where one would previously use `stale?` and
without the block is equivalent to `fresh_when?`.