SITES-701 ensure last_modified is included in strong ETag generation
By convention in a Rails controller the controller opts into what data
is included in an etag.  We want a site-wide policy that sets etags and
cache-control headers based on one or more records.

This fix ensures that the latest updated_at is considered as the
last_modified value even if the controller does not declare an etag.